requirements.txt
mwparserfromhell
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
discord.py==1.3.2
idna==2.8
multidict==4.5.2
praw==6.4.0
prawcore==1.0.1
pycparser==2.19
PyNaCl==1.3.0
qrcode==6.1
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.3
wavelink==0.3.4
websocket-client==0.56.0
websockets==6.0
yarl==1.3.0
youtube-dl==2020.3.24
